Consumer

NATS.JetStream.ConsumerConfiguration — Type

Configuration options for a consumer.

durable_name::Union{Nothing, String}: A unique name for a durable consumer

name::Union{Nothing, String}: A unique name for a consumer

description::Union{Nothing, String}: A short description of the purpose of this consumer

deliver_subject::Union{Nothing, String}

ack_policy::Symbol

ack_wait::Union{Nothing, Int64}: How long (in nanoseconds) to allow messages to remain un-acknowledged before attempting redelivery

max_deliver::Union{Nothing, Int64}: The number of times a message will be redelivered to consumers if not acknowledged in time

filter_subjects::Union{Nothing, Vector{String}}: Filter the stream by multiple subjects

replay_policy::Symbol

sample_freq::Union{Nothing, String}

rate_limit_bps::Union{Nothing, UInt64}: The rate at which messages will be delivered to clients, expressed in bit per second

max_ack_pending::Union{Nothing, Int64}: The maximum number of messages without acknowledgement that can be outstanding, once this limit is reached message delivery will be suspended

idle_heartbeat::Union{Nothing, Int64}: If the Consumer is idle for more than this many nano seconds a empty message with Status header 100 will be sent indicating the consumer is still alive

flow_control::Union{Nothing, Bool}: For push consumers this will regularly send an empty mess with Status header 100 and a reply subject, consumers must reply to these messages to control the rate of message delivery

max_waiting::Union{Nothing, Int64}: The number of pulls that can be outstanding on a pull consumer, pulls received after this is reached are ignored

headers_only::Union{Nothing, Bool}: Delivers only the headers of messages in the stream and not the bodies. Additionally adds Nats-Msg-Size header to indicate the size of the removed payload

max_batch::Union{Nothing, Int64}: The largest batch property that may be specified when doing a pull on a Pull Consumer

max_expires::Union{Nothing, Int64}: The maximum expires value that may be set when doing a pull on a Pull Consumer

max_bytes::Union{Nothing, Int64}: The maximum bytes value that maybe set when dong a pull on a Pull Consumer

inactive_threshold::Union{Nothing, Int64}: Duration that instructs the server to cleanup ephemeral consumers that are inactive for that long

backoff::Union{Nothing, Vector{Int64}}: List of durations in Go format that represents a retry time scale for NaK'd messages

num_replicas::Union{Nothing, Int64}: When set do not inherit the replica count from the stream but specifically set it to this amount

mem_storage::Union{Nothing, Bool}: Force the consumer state to be kept in memory rather than inherit the setting from the stream

Management

NATS.JetStream.consumer_create — Function

consumer_create(connection, config, stream)

Create a stream consumer.

NATS.JetStream.consumer_update — Function

consumer_update(connection, consumer, stream)

Update stream consumer configuration.

NATS.JetStream.consumer_delete — Function

consumer_delete(connection, stream_name, consumer_name)

Delete a consumer.

Subscriptions

NATS.JetStream.consumer_next — Function

consumer_next(
    connection,
    consumer,
    batch;
    no_wait,
    no_throw
)

Get next message for a consumer.

NATS.JetStream.consumer_ack — Function

consumer_ack(connection, msg; ...)
consumer_ack(connection, msg, ack; delays)

Confirms message delivery to server.