Request - reply

NATS.request — Function

request(nc, subject)
request(nc, subject, data; timer)

Send NATS Request-Reply message.

Default timeout is 5 seconds which can be overriden by passing timer.

Examples

julia> NATS.request(nc, "help.please")
NATS.Msg("l9dKWs86", "7Nsv5SZs", nothing, 17, "OK, I CAN HELP!!!")

julia> request(nc, "help.please"; timer = Timer(0))
ERROR: No replies received.

julia> request(nc, "help.please", nreplies = 2; timer = Timer(0))
NATS.Msg[]

request(subject, data, nreplies; timer)

The same as above but uses implicit connection (this functionality to be defined yet).

NATS.reply — Function

Reply for messages for a subject. Works like subscribe with automatic publish to the subject from reply_to field.

reply(f, nc, subject; queue_group, info)

Examples

julia> sub = reply("FOO.REQUESTS") do msg
    "This is a reply payload."
end
NATS.Sub("FOO.REQUESTS", nothing, "jdnMEcJN")

julia> sub = reply("FOO.REQUESTS") do msg
    "This is a reply payload.", ["example_header" => "This is a header value"]
end
NATS.Sub("FOO.REQUESTS", nothing, "jdnMEcJN")

julia> unsubscribe(sub)

The same as above but uses implicit connection (this functionality to be defined yet).

reply(f, subject; queue_group, info)

Examples

julia> nc = NATS.connect()
NATS.Connection(CONNECTED, 0 subs, 0 unsubs, 0 outbox)

julia> sub = reply(nc, "FOO.REQUESTS") do msg
    "This is a reply payload."
end