Connect

NATS.connect — Function

connect([host, port; options...])

Connect to NATS server. The function is blocking until connection is initialized.

Options are:

reconnect_delays: vector of delays that reconnect is performed until connected again, by default it will try to reconnect every second without time limit.

send_buffer_size: soft limit for buffer of messages pending. Default is 2097152 bytes, if too small operations that send messages to server (e.g. publish) may throw an exception

verbose: turns on protocol acknowledgements

pedantic: turns on additional strict format checking, e.g. for properly formed subjects

tls_required: indicates whether the client requires SSL connection

tls_ca_cert_path: CA certuficate file path

tls_client_cert_path: client public certificate file

tls_client_key_path: client private certificate file

auth_token: client authorization token

user: connection username

pass: connection password

name: client name

echo: if set to false, the server will not send originating messages from this connection to its own subscriptions

jwt: the JWT that identifies a user permissions and account.

no_responders: enable quick replies for cases where a request is sent to a topic with no responders.

nkey: the public NKey to authenticate the client

nkey_seed: the private NKey to authenticate the client

NATS.drain — Function

drain(nc)

Unsubscribe all subscriptions, wait for precessing all messages in buffers, then close connection. Drained connection is no more usable. This method is used to gracefuly stop the process.

drain()

drains all connections.