Request - reply

NATS.request — Function

request(subject)
request(subject, data; connection, timer)

Send NATS Request-Reply message.

Default timeout is 5.0 seconds which can be overriden by passing timer.

Optional keyword arguments are:

connection: connection to be used, if not specified default connection is taken

timer: error will be thrown if no replies received until timer expires

Examples

julia> NATS.request("help.please")
NATS.Msg("l9dKWs86", "7Nsv5SZs", nothing, 17, "OK, I CAN HELP!!!")

julia> request("help.please"; timer = Timer(0))
ERROR: No replies received.

request(subject, data, nreplies; connection, timer)

Requests for multiple replies. Vector of messages is returned after receiving nreplies replies or timer expired.

Optional keyword arguments are:

connection: connection to be used, if not specified default connection is taken

timer: error will be thrown if no replies received until timer expires

Examples

julia> request("help.please", nreplies = 2; timer = Timer(0))
NATS.Msg[]

NATS.reply — Function

reply(f, subject; connection, queue_group, async_handlers)

Reply for messages for a subject. Works like subscribe with automatic publish to the subject from reply_to field.

Optional keyword arguments are:

connection: connection to be used, if not specified default connection is taken

queue_group: NATS server will distribute messages across queue group members

async_handlers: if true task will be spawn for each f invocation, otherwise messages are processed sequentially, default is false

Examples

julia> sub = reply("FOO.REQUESTS") do msg
    "This is a reply payload."
end
NATS.Sub("FOO.REQUESTS", nothing, "jdnMEcJN")

julia> sub = reply("FOO.REQUESTS") do msg
    "This is a reply payload.", ["example_header" => "This is a header value"]
end
NATS.Sub("FOO.REQUESTS", nothing, "jdnMEcJN")

julia> unsubscribe(sub)